miterlimit is FIXED8 (float) SVG import - closing rectangle, stroke gradients

+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/commonshape/LineStyle.java
@@ -46,7 +46,7 @@ public class LineStyle {
 
     public boolean noClose;
 
-    public int miterLimitFactor;
+    public float miterLimitFactor;
 
     public boolean hasFillFlag;
